pyproject.toml
[project]
name = "answer-weaver"
version = "0.1.0"
description = "A bot that answers FAQs by referencing similar posts that have already been answered"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "praw>=7.8.1",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.1.0",
    "qdrant-client[fastembed]>=1.13.3",
    "sentence-transformers>=4.0.2",
    "uvicorn>=0.34.1",
]
